Introduce `CachedEventSource`
- `ValueTracker` implements `EventSource`
- `ValueTracker` implements `CachedEventSource`
