+++ src/index.ts
@@ -1,7 +1,9 @@
+export * from './cached-event-source';
 export * from './event-args';
 export * from './event-consumer';
 export * from './event-emitter';
 export * from './event-interest';
 export * from './event-producer';
+export * from './event-source';
 export * from './state';
 export * from './value';
