+++ src/value/index.ts
@@ -1,1 +1,2 @@
 export * from './value-tracker';
+export * from './tracked-value';
